+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 * Native data cache flushing implementation for `x86_64` and `AArch64`.
 * `large_slice` test case.
+* Benchmarking infrastructure using `Criterion.rs` to measure cache flush effectiveness.
 
 ### Changed
 * Switched focus from instruction cache synchronization to data cache flushing.